Covid-19 data for European countries

contains a dataframe with recorded daily deaths from Covid-19 in 11 European countries up until 05/05/2020. The dataframe includes variables representing different non-pharmaceutical interventions implemented by the countries considered. The data matches that used in Flaxman et al. (2020) . Also includes empirical distributions for the serial interval and the time from infection to death.

EuropeCovid

Format

A named list. The fields are:

data

A data frame giving indicators of certain non-pharmaceutical interventions in each country, along with death data and populations. The earliest date for each country in the dataframe is exactly 30 days before 10 cumulative deaths were observed in the country.

inf2death

A numeric vector representing the time distribution from infection to death assumed in Flaxman et al. (2020) .

si

The serial interval of covid-19 assumed in Flaxman et al. (2020) .

References

Flaxman S, Mishra S, Gandy A, Unwin HJT, Mellan TA, Coupland H, Whittaker C, Zhu H, Berah T, Eaton JW, Monod M, Perez-Guzman PN, Schmit N, Cilloni L, Ainslie KEC, Baguelin M, Boonyasiri A, Boyd O, Cattarino L, Cooper LV, Cucunubá Z, Cuomo-Dannenburg G, Dighe A, Djaafara B, Dorigatti I, van Elsland SL, FitzJohn RG, Gaythorpe KAM, Geidelberg L, Grassly NC, Green WD, Hallett T, Hamlet A, Hinsley W, Jeffrey B, Knock E, Laydon DJ, Nedjati-Gilani G, Nouvellet P, Parag KV, Siveroni I, Thompson HA, Verity R, Volz E, Walters CE, Wang H, Wang Y, Watson OJ, Winskill P, Xi X, Walker PGT, Ghani AC, Donnelly CA, Riley SM, Vollmer MAC, Ferguson NM, Okell LC, Bhatt S, Team ICCR (2020). “Estimating the effects of non-pharmaceutical interventions on COVID-19 in Europe.” Nature. ISSN 1476-4687, doi: 10.1038/s41586-020-2405-7 , https://doi.org/10.1038/s41586-020-2405-7.