Flexible Epidemic Modeling with epidemia

The epidemia package allows researchers to flexibly specify and fit Bayesian epidemiological models in the style of Flaxman et al. (2020) . The package leverages R's formula interface to parameterize the reproduction rate in terms of covariates, and allows pooling of parameters. The design of the package has been inspired by, and borrowed from, the rstanarm package (Goodrich et al. 2020) . epidemia uses cmdstanr as the backend for fitting the models, and the posterior package to represent posterior draws. The primary model fitting function in epidemia is epim.

References

Flaxman S, Mishra S, Gandy A, Unwin HJT, Mellan TA, Coupland H, Whittaker C, Zhu H, Berah T, Eaton JW, Monod M, Perez-Guzman PN, Schmit N, Cilloni L, Ainslie KEC, Baguelin M, Boonyasiri A, Boyd O, Cattarino L, Cooper LV, Cucunubá Z, Cuomo-Dannenburg G, Dighe A, Djaafara B, Dorigatti I, van Elsland SL, FitzJohn RG, Gaythorpe KAM, Geidelberg L, Grassly NC, Green WD, Hallett T, Hamlet A, Hinsley W, Jeffrey B, Knock E, Laydon DJ, Nedjati-Gilani G, Nouvellet P, Parag KV, Siveroni I, Thompson HA, Verity R, Volz E, Walters CE, Wang H, Wang Y, Watson OJ, Winskill P, Xi X, Walker PGT, Ghani AC, Donnelly CA, Riley SM, Vollmer MAC, Ferguson NM, Okell LC, Bhatt S, Team ICCR (2020). “Estimating the effects of non-pharmaceutical interventions on COVID-19 in Europe.” Nature. ISSN 1476-4687. doi:10.1038/s41586-020-2405-7 .

Goodrich B, Gabry J, Ali I, Brilleman S (2020). “rstanarm: Bayesian applied regression modeling via Stan.” https://mc-stan.org/rstanarm/. ()

See also

Author

Maintainer: Swapnil Mishra swapnilmishra.anu@gmail.com (ORCID)

Authors:

James Scott james.scott15@imperial.ac.uk (ORCID)

Axel Gandy a.gandy@imperial.ac.uk (ORCID)

Juliette Unwin h.unwin@imperial.ac.uk (ORCID)

Seth Flaxman s.flaxman@imperial.ac.uk (ORCID)

Samir Bhatt s.bhatt@imperial.ac.uk (ORCID)

Other contributors:

Jonathan Ish-Horowicz jonathan.ish-horowicz17@imperial.ac.uk (ORCID) [contributor]

Leon Wu leon.wu16@imperial.ac.uk [contributor]

Ben Goodrich benjamin.goodrich@columbia.edu [contributor]

Jonah Gabry jsg2201@columbia.edu [contributor]

Trustees of Columbia University (R/helpers.R) [copyright holder]

Paul-Christian Burkner paul.buerkner@gmail.com (R/misc.R) [copyright holder]

Triad Sou triadsou@gmail.com (R/geom_stepribbon.R) [copyright holder]

Kengo Nagashima (R/geom_stepribbon.R) [copyright holder]