Fit a Bayesian epidemiological model with epidemia

epim is the only model fitting function in epidemia. It takes a model description, a dataframe, and additional arguments relating to the fitting algorithm, and translates this to data that is then passed to a Stan program which is used to fit the model. The Stan program is compiled the first time it is needed and then cached (via cmdstanr), so subsequent calls to epim begin fitting immediately without recompilation.

epim(
  rt,
  inf,
  obs,
  data,
  algorithm = c("sampling", "meanfield", "fullrank"),
  group_subset = NULL,
  prior_PD = FALSE,
  ...
)

Arguments

rt

An object of class epirt. This defines the model for the time-varying reproduction rates. See epirt for more details.

inf

An object of class epiinf. This defines the model for latent infections. See epiinf for more details.

obs

Either an epiobs object, or a list of such objects. Each element in the list defines a model for the specified observation vector. See epiobs for more details.

data

A dataframe with all data required for fitting the model. This includes all observation variables and covariates specified in the models for the reproduction number and ascertainment rates.

algorithm

One of "sampling", "meanfield" or "fullrank". This specifies which cmdstanr fitting method to use. For "sampling" this is the $sample() method (adaptive HMC/NUTS); otherwise variational Bayes is used via the $variational() method.

group_subset

If specified, a character vector naming a subset of regions to include in the model.

prior_PD

If TRUE, samples all parameters from the joint prior distribution. This is useful for prior predictive checks. Defaults to FALSE.

...

Additional arguments (e.g. iter, chains, warmup, seed, cores, control) passed to the cmdstanr fitting method.

Value

An object of class epimodel.

Details

This is similar to the workflow for fitting Bayesian regression models with rstanarm. A key difference, however, is that the models fit by epidemia are much more complex, and are therefore inherently more difficult to specify. epidemia aims to simplify this process by modularizing the model definition into three distinct parts: transmission, infections and observations. These components of the model are defined with the functions epirt, epiinf and epiobs respectively.

epim has arguments rt, inf and obs which expect a description of the transmission model, infection model and observational models respectively. Together, these fully define the joint distribution of data and parameters. Each of these model components are described in terms of variables that are expected to live in a single dataframe, data. This dataframe must be compatible with the model components, in the sense that it holds all variables defined in these models.

In addition to taking a model description and a dataframe, epim has various additional arguments which specify how the model should be fit. If algorithm = "sampling" then the model will be fit using Stan’s adaptive Hamiltonian Monte Carlo sampler. This is done internally by calling the $sample() method of a cmdstanr model. If algorithm = "meanfield" or algorithm = "fullrank", then Stan’s variational Bayes algorithms are used instead, via the $variational() method. Any additional named arguments in the call to epim (for example iter, chains, warmup, seed, cores or control = list(adapt_delta=, max_treedepth=)) are translated and passed on to the underlying cmdstanr fitting method. epim returns a fitted model object of class epimodel, which contains posterior samples from the model along with other useful objects.

In general, the adaptive Hamiltonian Monte Carlo sampler should be used for final inference. Nonetheless, fitting these models using HMC is often computationally demanding, and variational Bayes can often be fruitful for quickly iterating models.

Examples

# Fitting requires a CmdStan installation (see cmdstanr::install_cmdstan()).
# \donttest{
library(EpiEstim)
data("Flu1918")

date <- as.Date("1918-01-01") + seq(0, along.with = c(NA, Flu1918$incidence))
data <- data.frame(
 city = "Baltimore",
 cases = c(NA, Flu1918$incidence),
 date = date,
 week = lubridate::week(date)
)

rt <- epirt(
 formula = R(city, date) ~ rw(time = week, prior_scale = 0.1),
 prior_intercept = normal(log(2), 0.2),
 link = 'log'
)

obs <-  epiobs(
 formula = cases ~ 1,
 prior_intercept = normal(location=1, scale=0.01),
 link = "identity",
 i2o = rep(.25,4)
)
#> Warning: the ‘findbars’ function has moved to the reformulas package. Please update your imports, or ask an upstream package maintainer to do so.

args <- list(
 rt = rt,
 inf = epiinf(gen = Flu1918$si_distr),
 obs = obs,
 data = data,
 algorithm = "sampling",
 iter = 1e3,
 seed = 12345
)

fm <- do.call(epim, args)
#> Running MCMC with 4 parallel chains...
#> 
#> Chain 1 Iteration:   1 / 1000 [  0%]  (Warmup) 
#> Chain 1 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 1 Exception: epidemia_base_model_namespace::log_prob: load[2, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 1 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 1 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 1 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 1 
#> Chain 1 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 1 Exception: epidemia_base_model_namespace::log_prob: load[2, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 1 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 1 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 1 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 1 
#> Chain 1 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 1 Exception: epidemia_base_model_namespace::log_prob: load[2, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 1 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 1 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 1 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 1 
#> Chain 1 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 1 Exception: epidemia_base_model_namespace::log_prob: load[12, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 1 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 1 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 1 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 1 
#> Chain 1 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 1 Exception: epidemia_base_model_namespace::log_prob: load[42, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 1 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 1 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 1 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 1 
#> Chain 1 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 1 Exception: epidemia_base_model_namespace::log_prob: load[2, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 1 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 1 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 1 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 1 
#> Chain 1 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 1 Exception: epidemia_base_model_namespace::log_prob: load[2, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 1 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 1 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 1 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 1 
#> Chain 1 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 1 Exception: epidemia_base_model_namespace::log_prob: load[14, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 1 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 1 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 1 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 1 
#> Chain 2 Rejecting initial value:
#> Chain 2   Error evaluating the log probability at the initial value.
#> Chain 2 Exception: neg_binomial_2_lpmf: Location parameter[1] is -2.00089e-07, but must be positive finite! (in '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 157, column 8 to line 158, column 61)
#> Chain 2 Exception: neg_binomial_2_lpmf: Location parameter[1] is -2.00089e-07, but must be positive finite! (in '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 157, column 8 to line 158, column 61)
#> Chain 2 Rejecting initial value:
#> Chain 2   Error evaluating the log probability at the initial value.
#> Chain 2 Exception: neg_binomial_2_lpmf: Location parameter[1] is -4.05733e-06, but must be positive finite! (in '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 157, column 8 to line 158, column 61)
#> Chain 2 Exception: neg_binomial_2_lpmf: Location parameter[1] is -4.05733e-06, but must be positive finite! (in '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 157, column 8 to line 158, column 61)
#> Chain 2 Rejecting initial value:
#> Chain 2   Error evaluating the log probability at the initial value.
#> Chain 2 Exception: neg_binomial_2_lpmf: Location parameter[1] is -6.59377e-06, but must be positive finite! (in '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 157, column 8 to line 158, column 61)
#> Chain 2 Exception: neg_binomial_2_lpmf: Location parameter[1] is -6.59377e-06, but must be positive finite! (in '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 157, column 8 to line 158, column 61)
#> Chain 2 Rejecting initial value:
#> Chain 2   Error evaluating the log probability at the initial value.
#> Chain 2 Exception: neg_binomial_2_lpmf: Location parameter[1] is -4.58991e-06, but must be positive finite! (in '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 157, column 8 to line 158, column 61)
#> Chain 2 Exception: neg_binomial_2_lpmf: Location parameter[1] is -4.58991e-06, but must be positive finite! (in '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 157, column 8 to line 158, column 61)
#> Chain 2 Rejecting initial value:
#> Chain 2   Error evaluating the log probability at the initial value.
#> Chain 2 Exception: neg_binomial_2_lpmf: Location parameter[1] is -3.37868e-06, but must be positive finite! (in '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 157, column 8 to line 158, column 61)
#> Chain 2 Exception: neg_binomial_2_lpmf: Location parameter[1] is -3.37868e-06, but must be positive finite! (in '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 157, column 8 to line 158, column 61)
#> Chain 2 Iteration:   1 / 1000 [  0%]  (Warmup) 
#> Chain 2 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 2 Exception: epidemia_base_model_namespace::log_prob: load[2, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 2 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 2 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 2 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 2 
#> Chain 2 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 2 Exception: epidemia_base_model_namespace::log_prob: load[2, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 2 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 2 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 2 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 2 
#> Chain 2 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 2 Exception: epidemia_base_model_namespace::log_prob: load[2, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 2 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 2 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 2 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 2 
#> Chain 2 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 2 Exception: epidemia_base_model_namespace::log_prob: load[14, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 2 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 2 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 2 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 2 
#> Chain 2 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 2 Exception: epidemia_base_model_namespace::log_prob: load[46, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 2 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 2 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 2 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 2 
#> Chain 2 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 2 Exception: epidemia_base_model_namespace::log_prob: load[2, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 2 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 2 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 2 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 2 
#> Chain 2 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 2 Exception: epidemia_base_model_namespace::log_prob: load[2, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 2 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 2 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 2 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 2 
#> Chain 2 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 2 Exception: epidemia_base_model_namespace::log_prob: load[15, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 2 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 2 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 2 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 2 
#> Chain 3 Rejecting initial value:
#> Chain 3   Error evaluating the log probability at the initial value.
#> Chain 3 Exception: neg_binomial_2_lpmf: Location parameter[1] is -6.54125e-06, but must be positive finite! (in '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 157, column 8 to line 158, column 61)
#> Chain 3 Exception: neg_binomial_2_lpmf: Location parameter[1] is -6.54125e-06, but must be positive finite! (in '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 157, column 8 to line 158, column 61)
#> Chain 3 Iteration:   1 / 1000 [  0%]  (Warmup) 
#> Chain 3 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 3 Exception: epidemia_base_model_namespace::log_prob: load[2, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 3 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 3 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 3 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 3 
#> Chain 3 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 3 Exception: epidemia_base_model_namespace::log_prob: load[2, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 3 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 3 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 3 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 3 
#> Chain 3 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 3 Exception: epidemia_base_model_namespace::log_prob: load[2, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 3 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 3 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 3 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 3 
#> Chain 3 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 3 Exception: epidemia_base_model_namespace::log_prob: load[14, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 3 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 3 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 3 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 3 
#> Chain 3 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 3 Exception: epidemia_base_model_namespace::log_prob: load[46, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 3 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 3 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 3 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 3 
#> Chain 3 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 3 Exception: epidemia_base_model_namespace::log_prob: load[2, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 3 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 3 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 3 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 3 
#> Chain 3 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 3 Exception: epidemia_base_model_namespace::log_prob: load[2, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 3 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 3 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 3 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 3 
#> Chain 3 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 3 Exception: epidemia_base_model_namespace::log_prob: load[16, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 3 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 3 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 3 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 3 
#> Chain 4 Rejecting initial value:
#> Chain 4   Error evaluating the log probability at the initial value.
#> Chain 4 Exception: neg_binomial_2_lpmf: Location parameter[1] is -1.7941e-06, but must be positive finite! (in '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 157, column 8 to line 158, column 61)
#> Chain 4 Exception: neg_binomial_2_lpmf: Location parameter[1] is -1.7941e-06, but must be positive finite! (in '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 157, column 8 to line 158, column 61)
#> Chain 4 Iteration:   1 / 1000 [  0%]  (Warmup) 
#> Chain 4 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 4 Exception: epidemia_base_model_namespace::log_prob: load[2, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 4 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 4 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 4 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 4 
#> Chain 4 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 4 Exception: epidemia_base_model_namespace::log_prob: load[2, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 4 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 4 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 4 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 4 
#> Chain 4 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 4 Exception: epidemia_base_model_namespace::log_prob: load[2, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 4 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 4 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 4 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 4 
#> Chain 4 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 4 Exception: epidemia_base_model_namespace::log_prob: load[14, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 4 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 4 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 4 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 4 
#> Chain 4 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 4 Exception: epidemia_base_model_namespace::log_prob: load[48, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 4 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 4 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 4 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 4 
#> Chain 4 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 4 Exception: epidemia_base_model_namespace::log_prob: load[2, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 4 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 4 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 4 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 4 
#> Chain 4 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 4 Exception: epidemia_base_model_namespace::log_prob: load[2, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 4 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 4 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 4 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 4 
#> Chain 4 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 4 Exception: epidemia_base_model_namespace::log_prob: load[16, 1] is -nan, but must be greater than or equal to 0.000000 (in '/home/runner/.cache/R/renv/library/epidemia2-11e6184b/linux-ubuntu-noble/R-4.6/x86_64-pc-linux-gnu/epidemia/stan//tparameters/infections_rt.stan', line 2, column 0, included from
#> Chain 4 '/tmp/RtmpgEpajl/model-67a0155d8028.stan', line 62, column 0)
#> Chain 4 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 4 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 4 
#> Chain 1 Iteration: 100 / 1000 [ 10%]  (Warmup) 
#> Chain 2 Iteration: 100 / 1000 [ 10%]  (Warmup) 
#> Chain 3 Iteration: 100 / 1000 [ 10%]  (Warmup) 
#> Chain 4 Iteration: 100 / 1000 [ 10%]  (Warmup) 
#> Chain 2 Iteration: 200 / 1000 [ 20%]  (Warmup) 
#> Chain 1 Iteration: 200 / 1000 [ 20%]  (Warmup) 
#> Chain 3 Iteration: 200 / 1000 [ 20%]  (Warmup) 
#> Chain 4 Iteration: 200 / 1000 [ 20%]  (Warmup) 
#> Chain 2 Iteration: 300 / 1000 [ 30%]  (Warmup) 
#> Chain 1 Iteration: 300 / 1000 [ 30%]  (Warmup) 
#> Chain 3 Iteration: 300 / 1000 [ 30%]  (Warmup) 
#> Chain 4 Iteration: 300 / 1000 [ 30%]  (Warmup) 
#> Chain 2 Iteration: 400 / 1000 [ 40%]  (Warmup) 
#> Chain 1 Iteration: 400 / 1000 [ 40%]  (Warmup) 
#> Chain 3 Iteration: 400 / 1000 [ 40%]  (Warmup) 
#> Chain 4 Iteration: 400 / 1000 [ 40%]  (Warmup) 
#> Chain 2 Iteration: 500 / 1000 [ 50%]  (Warmup) 
#> Chain 2 Iteration: 501 / 1000 [ 50%]  (Sampling) 
#> Chain 1 Iteration: 500 / 1000 [ 50%]  (Warmup) 
#> Chain 1 Iteration: 501 / 1000 [ 50%]  (Sampling) 
#> Chain 3 Iteration: 500 / 1000 [ 50%]  (Warmup) 
#> Chain 3 Iteration: 501 / 1000 [ 50%]  (Sampling) 
#> Chain 4 Iteration: 500 / 1000 [ 50%]  (Warmup) 
#> Chain 4 Iteration: 501 / 1000 [ 50%]  (Sampling) 
#> Chain 2 Iteration: 600 / 1000 [ 60%]  (Sampling) 
#> Chain 1 Iteration: 600 / 1000 [ 60%]  (Sampling) 
#> Chain 4 Iteration: 600 / 1000 [ 60%]  (Sampling) 
#> Chain 3 Iteration: 600 / 1000 [ 60%]  (Sampling) 
#> Chain 2 Iteration: 700 / 1000 [ 70%]  (Sampling) 
#> Chain 4 Iteration: 700 / 1000 [ 70%]  (Sampling) 
#> Chain 3 Iteration: 700 / 1000 [ 70%]  (Sampling) 
#> Chain 1 Iteration: 700 / 1000 [ 70%]  (Sampling) 
#> Chain 2 Iteration: 800 / 1000 [ 80%]  (Sampling) 
#> Chain 4 Iteration: 800 / 1000 [ 80%]  (Sampling) 
#> Chain 3 Iteration: 800 / 1000 [ 80%]  (Sampling) 
#> Chain 2 Iteration: 900 / 1000 [ 90%]  (Sampling) 
#> Chain 1 Iteration: 800 / 1000 [ 80%]  (Sampling) 
#> Chain 4 Iteration: 900 / 1000 [ 90%]  (Sampling) 
#> Chain 2 Iteration: 1000 / 1000 [100%]  (Sampling) 
#> Chain 2 finished in 31.7 seconds.
#> Chain 3 Iteration: 900 / 1000 [ 90%]  (Sampling) 
#> Chain 1 Iteration: 900 / 1000 [ 90%]  (Sampling) 
#> Chain 3 Iteration: 1000 / 1000 [100%]  (Sampling) 
#> Chain 4 Iteration: 1000 / 1000 [100%]  (Sampling) 
#> Chain 3 finished in 34.0 seconds.
#> Chain 4 finished in 34.0 seconds.
#> Chain 1 Iteration: 1000 / 1000 [100%]  (Sampling) 
#> Chain 1 finished in 35.8 seconds.
#> 
#> All 4 chains finished successfully.
#> Mean chain execution time: 33.9 seconds.
#> Total execution time: 35.9 seconds.
#> 
#> Warning: 72 of 2000 (4.0%) transitions hit the maximum treedepth limit of 10.
#> See https://mc-stan.org/misc/warnings for details.

# Variational Bayes ("meanfield" / "fullrank") is much faster and useful for
# iterating on a model, but it is not reliable for every model: ADVI can fail
# to compute an initial ELBO, and it does not approximate the Gamma priors
# used for NPI effects well. Use "sampling" for anything you intend to report.

# }