Helper for constructing an object of class 'epirt'

Defines a model for the latent time varying reproduction number. For more details on the model assumptions please refer to the online vignettes.

epirt(
  formula,
  r0 = 3.28,
  center = FALSE,
  prior = rstanarm::normal(scale = 0.5),
  prior_intercept = rstanarm::normal(scale = 0.5),
  prior_covariance = rstanarm::decov(scale = 0.5),
  ...
)

Arguments

formula
An R object of class "formula". The left hand side must take the form R(group,date), with group representing a factor vector indicating group membership (i.e. country, state, age cohort), and date being a vector of Date objects.

r0
The prior expected value of \(R_0\). The maximum \(R_0\) in the simulations will be limited to twice this value.

center
If TRUE then the covariates for the \(R_t\) regression are centered to have mean zero. All of the priors are then interpreted as prior on the centered covariates. Defaults to FALSE.

prior
Same as in stan_glm. In addition to the rstanarm provided priors, a shifted_gamma can be used. Note: If autoscale=TRUE in the call to the prior distribution then automatic rescaling of the prior may take place.

prior_intercept
Same as in stan_glm. Prior for the regression intercept (if it exists).

prior_covariance
Same as in stan_glmer. Only used if the formula argument specifies random effects.

...
Additional arguments for model.frame