Posterior model evaluations

Calculate daily error using one of three metrics, and also return coverage of credible intervals. Uses continuous ranked probability score (CRPS), mean absolute error and median absolute error.

evaluate_forecast(
  object,
  newdata = NULL,
  type,
  groups = NULL,
  metrics = NULL,
  levels = c(50, 95)
)

Arguments

object
A fitted model object returned by epim. See epimodel-objects.

newdata
If provided, the original data used in object is overridden. Useful for forecasting

type
A string specifying the name of the observations to plot. This should match one of the names of the response variables in the obs argument used int the call to epim.

groups
Either NULL or a character vector specifying the groups to plot for. Default is NULL, which plots all modeled groups.

metrics
A string or character vector specifying the plotted forecast error metrics. One of NULL, "crps", "mean_abs_error"

levels
A numeric vector defining the levels of the plotted credible intervals.

Value

A named list with dataframes giving metrics and coverage.