Plotting the posterior predictive distribution

Plots credible intervals for the observed data under the posterior predictive distribution. Plots for a specific observation type. The user can control the levels of the intervals and the plotted group(s). This is a generic function.

plot_obs(object, ...)

# S3 method for epimodel
plot_obs(
  object,
  type = NULL,
  posterior_mean = FALSE,
  group = NULL,
  dates = NULL,
  date_breaks = "2 weeks",
  date_format = "%Y-%m-%d",
  cumulative = FALSE,
  levels = c(50, 95),
  log = FALSE,
  ...
)

Arguments

object
A fitted model object returned by epim. See epimodel-objects.

...
Additional arguments for posterior_predict. Examples include newdata, which allows predictions or counterfactuals.

type
the name of the observations to plot. This should match one of the names of the obs argument to epim.

posterior_mean
If true, the credible intervals are plotted for the posterior mean. Defaults to FALSE, in which case the posterior predictive is plotted.

group
NULL, string or character vector specifying which groups to plot. Default is NULL, which plots all possible groups.

dates
a vector of (start date, end date) defining the date range to be plotted. Must be coercible to date if not NA. If NA, this means use the lower/upper limit as appropriate. See examples.

date_breaks
string giving the distance between date tick labels. Default is "2 weeks". Passed as date_breaks argument to ggplot::scale_x_date - see https://ggplot2.tidyverse.org/reference/scale_date.html

date_format
the date format for coercing dates. Default is "%Y-%m-%d"

cumulative
If TRUE, plots the cumulative observations. Defaults to FALSE

levels
numeric vector giving the levels of the plotted credible intervals

log
If TRUE, plots the observations on a pseudo-linear scale. Defaults to FALSE.

Value

A ggplot object which can be further modified.

Examples

if (FALSE) {
## load required data
library(epidemia)
data("EuropeCovid")
## setup sampling
args <- EuropeCovid
args$algorithm <- "sampling"
args$sampling_args <- list(iter=1e3,control=list(adapt_delta=0.95,max_treedepth=15),seed=12345)
args$group_subset <- c("Italy")
args$formula <- R(country,date) ~  1 + lockdown
args$prior <- rstanarm::normal(location=0,scale=.5)
args$prior_intercept <- rstanarm::normal(location=0,scale=2)

## run sampling
fit <- do.call("epim", args)

## make plots
plot_obs(fit, type="deaths") # default, plots all groups and dates
plot_obs(fit, type="deaths", dates=c("2020-03-21", NA)) # plot 21 March 2020 onwards
plot_obs(fit, type="deaths", dates=c(NA, "2020-03-20")) # plot up to  20 March 2020
plot_obs(fit, type="deaths", dates=c("2020-03-20", "2020-04-20")) # plot 20 March-20 April 2020
plot_obs(fit, type="deaths", dates=c("2020-03-20", "2020-04-20"), date_breaks="1 day") # plot 21 March-20 April 2020 with ticks every day
plot_obs(fit, type="deaths", dates=c("2020-03-20", "2020-04-20"), date_breaks="1 week") # plot 21 March-20 April 2020 with ticks every week
plot_obs(fit, type="deaths", dates=c("2020-20-03", "2020-20-04"), date_format="%Y-%d-%m") # plot 21 March-20 April 2020 (different date format)
}