Gives the posterior linear predictor for the reproduction numbers Will be extended for observations in future versions

Gives the posterior linear predictor for the reproduction numbers Will be extended for observations in future versions

posterior_linpred(
  object,
  transform = FALSE,
  type = NULL,
  newdata = NULL,
  draws = NULL,
  fixed = TRUE,
  random = TRUE,
  autocor = TRUE,
  offset = TRUE,
  ...
)

Arguments

object
A fitted model object returned by epim. See epimodel-objects.

transform
If TRUE, transforms the predictor by the inverse link function. Defaults to FALSE.

type
If NULL, then gives posterior linear predictor for reproduction numbers. Otherwise gives the predictor for the specified observation type.

newdata
If provided, the original data used in object is overidden. Useful both for conterfactual and prediction analysis

draws
Number of posterior draws to use. Defaults to the number of parameter draws in the fitted model.

fixed
Include fixed effects. Defaults to TRUE.

random
Include random effects. Defaults to TRUE.

autocor
Include autocorrelation terms. Defaults to TRUE.

offset
Include offsets. Defaults to TRUE.

...
Not used.