A shifted gamma prior

A gamma prior distribution which can be shifted.

shifted_gamma(shape = 1, scale = 1, shift = 0, autoscale = TRUE)

Arguments

shape, scale

Sets the shape and scale parameters of the Gamma prior.

shift

The Gamma prior can be shifted to allow for positive support.

autoscale

Same as in priors.

Value

A named list to be parsed internally by epim.

Details

rstanarm provides a set of distributions (see priors) which can be used for the priors on regression parameters. Intuitively, non-pharmaceutical interventions are unlikely to a-priori cause a large increase in the reproduction number. A shifted gamma prior can be used to model this idea, and has been used in Flaxman et al. (2020) . shifted_gamma can be used as the prior argument to epim. This specified a shifted gamma prior on the negative of the regression parameters. i.e. if there is no shift, the support is on the negative half of the real line.

References

Flaxman S, Mishra S, Gandy A, Unwin HJT, Mellan TA, Coupland H, Whittaker C, Zhu H, Berah T, Eaton JW, Monod M, Perez-Guzman PN, Schmit N, Cilloni L, Ainslie KEC, Baguelin M, Boonyasiri A, Boyd O, Cattarino L, Cooper LV, Cucunubá Z, Cuomo-Dannenburg G, Dighe A, Djaafara B, Dorigatti I, van Elsland SL, FitzJohn RG, Gaythorpe KAM, Geidelberg L, Grassly NC, Green WD, Hallett T, Hamlet A, Hinsley W, Jeffrey B, Knock E, Laydon DJ, Nedjati-Gilani G, Nouvellet P, Parag KV, Siveroni I, Thompson HA, Verity R, Volz E, Walters CE, Wang H, Wang Y, Watson OJ, Winskill P, Xi X, Walker PGT, Ghani AC, Donnelly CA, Riley SM, Vollmer MAC, Ferguson NM, Okell LC, Bhatt S, Team ICCR (2020). “Estimating the effects of non-pharmaceutical interventions on COVID-19 in Europe.” Nature. ISSN 1476-4687. doi:10.1038/s41586-020-2405-7 .