Flexible Epidemic Modeling with epidemia

The epidemia package allows researchers to flexibly specify and fit Bayesian epidemiological models in the style of Flaxman et al. (2020) . The package leverages R's formula interface to paramterise the reproduction rate in terms of covariates, and allows pooling of parameters. The design of the package has been inspired by, and borrowed from, the rstanarm package (Goodrich et al. 2020) . epidemia uses rstan (Stan Development Team 2020) as the backend for fitting the models. The primary model fitting function in epidemia is epim.

References

Flaxman S, Mishra S, Gandy A, Unwin HJT, Mellan TA, Coupland H, Whittaker C, Zhu H, Berah T, Eaton JW, Monod M, Perez-Guzman PN, Schmit N, Cilloni L, Ainslie KEC, Baguelin M, Boonyasiri A, Boyd O, Cattarino L, Cooper LV, Cucunubá Z, Cuomo-Dannenburg G, Dighe A, Djaafara B, Dorigatti I, van Elsland SL, FitzJohn RG, Gaythorpe KAM, Geidelberg L, Grassly NC, Green WD, Hallett T, Hamlet A, Hinsley W, Jeffrey B, Knock E, Laydon DJ, Nedjati-Gilani G, Nouvellet P, Parag KV, Siveroni I, Thompson HA, Verity R, Volz E, Walters CE, Wang H, Wang Y, Watson OJ, Winskill P, Xi X, Walker PGT, Ghani AC, Donnelly CA, Riley SM, Vollmer MAC, Ferguson NM, Okell LC, Bhatt S, Team ICCR (2020). “Estimating the effects of non-pharmaceutical interventions on COVID-19 in Europe.” Nature. ISSN 1476-4687, doi: 10.1038/s41586-020-2405-7 , https://doi.org/10.1038/s41586-020-2405-7.

Goodrich B, Gabry J, Ali I, Brilleman S (2020). “rstanarm: Bayesian applied regression modeling via Stan.” R package version 2.19.3, https://mc-stan.org/rstanarm.

Stan Development Team (2020). “RStan: the R interface to Stan.” R package version 2.19.3, http://mc-stan.org/. () Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org