Model Latent Infections

epiinf defines a model for latent infections. For the basic version of the model, this defines the generation distribution of the disease, the number of days for which to seed infections, and the prior distribution on the parameter \(\tau\), as described in the model description vignette. Recall that \(\tau\) is the prior mean on daily seeded infections. These three parameters are controlled by the arguments gen, seed_days and prior_tau respectively.

epiinf(
  gen,
  seed_days = 6L,
  latent = FALSE,
  family = "log-normal",
  prior_aux = rstanarm::normal(10, 5),
  prior_tau = rstanarm::exponential(0.03),
  pop_adjust = FALSE,
  susceptibles = NULL
)

Arguments

gen
A numeric vector representing the probability mass function for the generation time of the disease (must be a probability vector).

seed_days
An integer giving the number of days for which to seed infections. Defaults to 6L.

latent
If TRUE, treat infections as latent parameters using the extensions described in Section 5.2 here.

family
Specifies the family for the prior distribution on daily infections. Only used if latent = TRUE, and currently restricted to log-normal.

prior_aux
Prior distribution for the coefficient of dispersion \(d\) for the offspring distribution. Only used if latent = TRUE. The number of offspring of a given infection is assumed to have mean \(\mu\) and variance \(d \mu\). This argument specifies prior on \(d\). Higher values of \(d\) imply more super-spreading events.

prior_tau
The prior distribution for the hyperparameter \(\tau\), which is the mean of the prior distribution for infection seeding. Must be a call to exponential. Defaults to rstanarm::exponential(0.03). See the model description for more details on this parameter.

pop_adjust
If TRUE, applies a population adjustment to the infection process. Defaults to FALSE.

susceptibles
A character vector giving the name of the column in the dataframe passed as the data argument of epim, that corresponds to the susceptible population over time. Only used if pop_adjust=TRUE.

Value

An object of class epiinf.

Details

epiinf has additional arguments which allow the user to extend the basic model. Using latent=TRUE replaces the renewal process with a model that treats latent infections as unknown parameters that are sampled along with other parameters. The family argument then gives the distribution family for latent infections, while prior_aux defines the prior on the coefficient of dispersion \(d\) of this distribution.

Recall that one can adjust the infection process to explicitly model changes in infection rates as the remaining susceptible population is depleted. In particular, the adjustment ensures that cumulative infections never breaches the initial susceptible population. The adjustment was described in Section 5.3 of the model description article. It can be employed by setting pop_adjust = TRUE and using the susceptibles argument to point towards a variable in the dataframe which gives the susceptible population at each point in time.

Examples

data(EuropeCovid)
inf <- epiinf(
 gen = EuropeCovid$si,
 seed_days = 6L,
 prior_tau = rstanarm::exponential(0.02)
)