Fits an Epidemiological Model

Fits a Bayesian epidemiological model specified by the formula argument.

epim(
  rt,
  inf,
  obs,
  data,
  algorithm = c("sampling", "meanfield", "fullrank"),
  group_subset = NULL,
  prior_PD = FALSE,
  init_run = FALSE,
  ...
)

Arguments

rt
An object of class epirt. This specifies the model for the time-varying reproduction number.

inf
An object of class epiinf. This specifies the model for latent infections. Defaults to epiinf().

obs
A list of epiobs objects. Each element defines a model for the specified observation vector.

data
A dataframe containing all data required to fit the model. See lm.

algorithm
One of "sampling", "meanfield" or "fullrank". This specifies which rstan function to use for fitting the model.

group_subset
An optional vector specifying a subset of groups to model. Elements should correspond to the group levels specified through the data argument.

prior_PD
Same as in stan_glm. If TRUE, samples parameters from the prior disribution. Defaults to FALSE.

init_run
For certain datasets the sampler can find itself trapped in a local mode where herd immunity is achieved. If TRUE, an MCMC run where the population adjustment is disabled is used to initialise the parameters for the main sampling. If TRUE, this is done with default parameters. If instead a list is provided, these parameters are passed on to the rstan sampling function for the initial run. initial run. The seed used is that specified in init_run, or that specified in ..., or no seed, in that order.

...
An (optional) named list of parameters to pass to the rstan function used for model fitting, for example rstan::sampling.

Value

An object of class "epimodel".

Details

epim is the primary model fitting function in epidemia, and fits models in the style of Flaxman et al. (2020) . Multiple groups (countries/states/age cohorts) can be modeled simultaneously using multilevel models. The time-varying reproduction number can be paramterised by a number of covariates through the formula argument. This is reasonably flexible. For example, random effects terms in the style of the lme4 package can be included. The prior distribution on the parameters in the regression are handled through the arguments prior, prior_intercept and prior_covariance.

References

Flaxman S, Mishra S, Gandy A, Unwin HJT, Mellan TA, Coupland H, Whittaker C, Zhu H, Berah T, Eaton JW, Monod M, Perez-Guzman PN, Schmit N, Cilloni L, Ainslie KEC, Baguelin M, Boonyasiri A, Boyd O, Cattarino L, Cooper LV, Cucunubá Z, Cuomo-Dannenburg G, Dighe A, Djaafara B, Dorigatti I, van Elsland SL, FitzJohn RG, Gaythorpe KAM, Geidelberg L, Grassly NC, Green WD, Hallett T, Hamlet A, Hinsley W, Jeffrey B, Knock E, Laydon DJ, Nedjati-Gilani G, Nouvellet P, Parag KV, Siveroni I, Thompson HA, Verity R, Volz E, Walters CE, Wang H, Wang Y, Watson OJ, Winskill P, Xi X, Walker PGT, Ghani AC, Donnelly CA, Riley SM, Vollmer MAC, Ferguson NM, Okell LC, Bhatt S, Team ICCR (2020). “Estimating the effects of non-pharmaceutical interventions on COVID-19 in Europe.” Nature. ISSN 1476-4687, doi: 10.1038/s41586-020-2405-7 , https://doi.org/10.1038/s41586-020-2405-7.

Examples

if (FALSE) {
data("EuropeCovid")

args <- EuropeCovid
args$algorithm <- "sampling"
args$rt <- epirt(
 formula=R(country, date) ~ 0 + lockdown,
 args$prior <- shifted_gamma(shape = 1 / 6,
                             scale = 1,
                             shift = log(1.05) / 6)
)

fit <- do.call("epim", args)
plot_rt(fit, group = "Germany")
}