Fits an Epidemiological Model

Fits a Bayesian epidemiological model specified by the formula argument.

epim(
  formula,
  data,
  obs,
  pops,
  si,
  seed_days = 6,
  algorithm = c("sampling", "meanfield", "fullrank"),
  group_subset = NULL,
  stan_data = FALSE,
  center = FALSE,
  prior = rstanarm::normal(scale = 0.5),
  prior_intercept = rstanarm::normal(scale = 0.5),
  prior_covariance = rstanarm::decov(scale = 0.5),
  r0 = 3.28,
  prior_phi = rstanarm::normal(location = 0, scale = 5),
  prior_tau = rstanarm::exponential(rate = 0.03),
  prior_PD = FALSE,
  sampling_args = list(),
  debug = FALSE,
  ...
)

Arguments

formula
An R object of class "formula". The left hand side must take the form R(group,date), with group representing a factor vector indicating group membership (i.e. country, state, age cohort), and date being a vector of Date objects.

data
A dataframe with columns corresponding to the terms appearing in 'formula'. See lm.

obs
A list of lists giving available observations. Each element of 'obs' must itself be a names list containing the following three elements

odata: A three column dataframe representing some type of observed data that is a function of the true number of infections. Examples include recorded incidence, deaths or hospitalisations. The first column represents group membership and must be coercible to class factor. The second column indicates the observation date and must be coercible to class Date. The third column contain the data.

rates: A named list specifying the prior for the proportion of infected individuals recorded as an observation. For example if recording deaths this would be some estimate of each groups infection fatality ratio (IFR). The priors are assumed normal. Contains:

means: A two column dataframe giving the estimated mean proportion of infected individuals recorded as an observation. First column represents the group, while the second the corresponding proportion.

scale: The standard deviation around the mean values. Default value is 0.1.

pvec: A probability vector with the following interpretation. Conditional on an observation "event" (i.e. a single death or hospitalisation etc.), the nth element represents the probability that the individual was infected exactly n days prior to this.

ptype: A string, either "density", "distribution" or "unadjusted". Indicates how pvec should be treated. If "density" then rescaled to simplex vector, If "distribution", ensures it takes the form of a distribution function. If "unadjusted", only checks values are between 0 and 1. simplex vector. Otherwise, this is treated as a cumulative distribution. Defaults to TRUE.

pops
A two column dataframe giving the total population of each group. First column represents the group, with the second giving the corresponding population.

si
A vector representing the serial interval of the disease (a probability vector).

seed_days
Number of days for which to seed infections.

algorithm
One of "sampling", "meanfield" or "fullrank". This specifies which rstan function to use for fitting the model.

group_subset
An optional vector specifying a subset of groups to model. Elements should correspond to the group levels specified through the data argument.

stan_data, debug
Mainly for internal use. Will be removed in future versions.

center
If TRUE then the covariates for the \(R_t\) regression are centered to have mean zero. All of the priors are then interpreted as prior on the centered covariates. Defaults to FALSE.

prior
Same as in stan_glm. In addition to the rstanarm provided priors, a shifted_gamma can be used. Note: If autoscale=TRUE (Default) in the call to the prior distribution then automatic rescaling of the prior may take place.

prior_intercept
Same as in stan_glm. Prior for the regression intercept (if it exists).

prior_covariance
Same as in stan_glmer. Only used if the formula argument specifies a

r0
The prior expected value of \(R_0\). The maximum \(R_0\) in the simulations will be limited to twice this value.

prior_phi
The prior distribution on \(\phi\). This parameter is described in the introductory vignette, and determined the variance of the observed data around its mean. Must be a call to normal, which again is transformed to a half normal distribution.

prior_tau
The prior for \(\tau\).This parameter is described in the introductory vignette, and controls the variability in the number of seeded infections at the beginning of the epidemic. Must be a call to exponential.

prior_PD
Same as in stan_glm. If TRUE, samples parameters from the prior disribution. Defaults to FALSE.

sampling_args
An (optional) named list of parameters to pass to the rstan function used for model fitting, for example rstan::sampling.

...
Not used.

Value

An object of class "epimodel".

Details

epim is the primary model fitting function in epidemia, and fits models in the style of Flaxman et al. (2020) . Multiple groups (countries/states/age cohorts) can be modeled simultaneously using multilevel models. The time-varying reproduction number can be paramterised by a number of covariates through the formula argument. This is reasonably flexible. For example, random effects terms in the style of the lme4 package can be included. The prior distribution on the parameters in the regression are handled through the arguments prior, prior_intercept and prior_covariance.

References

Flaxman S, Mishra S, Gandy A, Unwin HJT, Mellan TA, Coupland H, Whittaker C, Zhu H, Berah T, Eaton JW, Monod M, Perez-Guzman PN, Schmit N, Cilloni L, Ainslie KEC, Baguelin M, Boonyasiri A, Boyd O, Cattarino L, Cooper LV, Cucunubá Z, Cuomo-Dannenburg G, Dighe A, Djaafara B, Dorigatti I, van Elsland SL, FitzJohn RG, Gaythorpe KAM, Geidelberg L, Grassly NC, Green WD, Hallett T, Hamlet A, Hinsley W, Jeffrey B, Knock E, Laydon DJ, Nedjati-Gilani G, Nouvellet P, Parag KV, Siveroni I, Thompson HA, Verity R, Volz E, Walters CE, Wang H, Wang Y, Watson OJ, Winskill P, Xi X, Walker PGT, Ghani AC, Donnelly CA, Riley SM, Vollmer MAC, Ferguson NM, Okell LC, Bhatt S, Team ICCR (2020). “Estimating the effects of non-pharmaceutical interventions on COVID-19 in Europe.” Nature. ISSN 1476-4687, doi: 10.1038/s41586-020-2405-7 , https://doi.org/10.1038/s41586-020-2405-7.

Examples

if (FALSE) {
data("EuropeCovid")

args <- EuropeCovid
args$algorithm <- "sampling"
args$formula <- R(country,date) ~ 0 + lockdown
args$prior <- shifted_gamma(shape = 1/6, scale = 1, shift = -log(1.05)/6)

fit <- do.call("epim", args)
plot_rt(fit, group = "Germany")
}