Helper for constructing an object of class 'epiobs'

Defines a model for an observation vector. These observations are taken to be a function of the latent infections in the poulation. Examples include daily death or hospitalisation rates. For more details on the model assumptions please refer to the online vignettes.

epiobs(
  formula,
  family = "neg_binom",
  link = "logit",
  i2o,
  center = FALSE,
  offset = NULL,
  prior = rstanarm::normal(scale = 0.1),
  prior_intercept = rstanarm::normal(scale = 0.1),
  prior_aux = rstanarm::exponential(autoscale = TRUE),
  ...
)

Arguments

formula
A formula defining the model for the observations.

family
A string representing the error distribution for the model. Can be either "poisson" or "neg_binom".

link
A string representing the link function used to transform the covariates. The linear predictor constructed from the covariates is transformed by the inverse link function, then multiplied by the weighted previous infections. This quantity represents the mean of the response distribution.

i2o
A probability vector with the following interpretation. Conditional on an observation "event" (i.e. a single death or hospitalisation etc.), the nth element represents the probability that the individual was infected exactly n days prior to this.

center
If TRUE then the covariates are centered to have mean zero. All of the priors are then interpreted as priors on the centered covariates. Defaults to FALSE.

offset
Same as glm

prior
Same as in stan_glm. Note: If autoscale=TRUE in the call to the prior distribution then automatic rescaling of the prior may take place.

prior_intercept
Same as in stan_glm. Prior for the regression intercept, if one has been specified.

prior_aux
Specify the prior distribution for the auxiliary parameter if it exists. Only used if family is negative binomial, in which case this represents the prior on the reciprocal of the dispersion parameter. See stan_glm for more details.

...
Additional arguments for model.frame