Plotting the posterior linear predictor for R or ascertainments

Plots credible intervals for the observed data under the posterior predictive distribution, and for a specific observation type. The user can control the levels of the intervals and the plotted group(s). This is a generic function.

plot_linpred(object, ...)

# S3 method for epimodel
plot_linpred(
  object,
  type = NULL,
  groups = NULL,
  dates = NULL,
  date_breaks = "2 weeks",
  date_format = "%Y-%m-%d",
  levels = c(30, 60, 90),
  plotly = FALSE,
  ...
)

Arguments

object
A fitted model object returned by epim. See epimodel-objects.

...
Additional arguments for posterior_predict.epimodel. Examples include newdata, which allows predictions or counterfactuals.

type
the name of the observations to plot. This should match one of the names of the obs argument to epim.

groups
NULL, string or character vector specifying which groups to plot. Default is NULL, which plots all possible groups.

dates
a vector of (start date, end date) defining the date range to be plotted. Must be coercible to date if not NA. If NA, this means use the lower/upper limit as appropriate. See examples.

date_breaks
string giving the distance between date tick labels. Default is "2 weeks". Passed as date_breaks argument to ggplot::scale_x_date - see https://ggplot2.tidyverse.org/reference/scale_date.html

date_format
the date format for coercing dates. Default is "%Y-%m-%d"

levels
numeric vector giving the levels of the plotted credible intervals

plotly
If TRUE, wraps the ggplot object into a plotly object, for interactive graphing.

Value

A ggplot object which can be further modified.